+++ CHANGELOG.md
@@ -5,5 +5,8 @@ Items:
   - 'Item:spawner, Glowing:true'
 ```
 
+### Fixed:
+- The config option for `verbose-logging` was not applied to some parts of the plugin.
+
 ### Changes:
 - No longer add the contents of `DisplayData` to the `Items` section on `/crazycrates reload`